Made PostComment into smaller more manageble components CommentForm, CommentItem, CommentList, and CommentSection. And deleted PostComment
